Major Structural Improvements Fixed the bugs with the webcam not working on some PCs.
