+++ Registration Application/build.gradle
@@ -100,11 +100,16 @@ dependencies {
     compile 'com.squareup.dagger:dagger:1.2.2'
     compile 'commons-configuration:commons-configuration:1.10'
 
-    //compile group: 'org.bytedeco', name: 'javacv', version: '0.11'
-    //compile group: 'org.bytedeco', name: 'javacpp', version: '0.11'
-    //compile 'org.bytedeco.javacpp-presets:opencv:2.4.11-0.11'
+
+    compile 'org.bytedeco:javacv:1.0'
+    compile 'org.bytedeco.javacpp-presets:opencv:3.0.0-1.0:windows-x86_64'
+    compile 'org.bytedeco.javacpp-presets:ffmpeg:2.7.1-1.0:windows-x86_64'
+
+
+
+
     compile 'com.google.code.gson:gson:2.3.1'
-    //compile 'org.bytedeco.javacpp-presets:ffmpeg:2.6.1-0.11'
+    compile 'org.bytedeco.javacpp-presets:ffmpeg:2.6.1-0.11'
     //compile files("libs/opencv-2411.jar")
     compile 'commons-configuration:commons-configuration:1.10'
     compile 'com.googlecode.json-simple:json-simple:1.1.1'
